Lots of improvements to image generation service: - Add pure Java2D implementation that handles transparent PNG well - Java2D is also used by default for JPEG and GIF formats. It does not support JPEG2000 or CMYK JPEGs though. Transparent GIFs should work except for rotation, where some banding occurs for some unknown reason (JVM bug ?) - Improve error handling to only fail on the thumbnails that cannot be read instead of stopping the process completely. - Improved the ImageJ library to handle transparent TIFFs better, although transparency is not preserved. - The new default implementation ImageJAndJava2DImageService actually uses both implementations : Java2D for JPEG, GIF and PNGs images, and ImageJ for all other formats. - Fixed some bugs in ImageMagick implementation that was causing cropped images to have an improper target size (same bug for square image resize), as well as error handling to make sure that only images not supported fail instead of the whole thumbnail generation rule. - A lot of testing of these libraries was done by creating an external test project, that can be found here : https://github.com/sergehuber/ImageSupportTests . This external test project also made it possible to analyze the performance of the various image generation sub-systems. - Also improved the naming of Jahia-generated temporary files to include the proper extension for the temporary files, as this was confusing some image generation libraries such as ImageMagick. - Added lots of JavaDocs to JahiaImageService interface (it had none :)) (merged from jahia/branches/JAHIA-6-6-X-X-BRANCH revision 41448)
git-svn-id: svn://localhost/jahia/trunk@16214 69c369ed-ed09-0410-a4a1-e19a46c23626

+++ core/src/main/java/org/jahia/services/content/rules/ImageService.java
@@ -40,6 +40,7 @@
 
 package org.jahia.services.content.rules;
 
+import org.apache.commons.io.FilenameUtils;
 import org.drools.ObjectFilter;
 import org.drools.spi.KnowledgeHelper;
 import org.jahia.api.Constants;
@@ -95,6 +96,9 @@ public class ImageService {
             return (Image) it.next();
         }
         Image iw = imageService.getImage(imageNode.getNode());
+        if (iw == null) {
+            return null;
+        }
         drools.insertLogical(iw);
         return iw;
     }
@@ -110,11 +114,17 @@ public class ImageService {
             if (contentDate.after(thumbDate)) {
                 AddedNodeFact thumbNode = new AddedNodeFact(node);
                 File f = getThumbFile(imageNode, size,square, drools);
+                if (f == null) {
+                    return;
+                }
                 drools.insert(new ChangedPropertyFact(thumbNode, Constants.JCR_DATA, f, drools));
                 drools.insert(new ChangedPropertyFact(thumbNode, Constants.JCR_LASTMODIFIED, new GregorianCalendar(), drools));
             }
         } else {
             File f = getThumbFile(imageNode, size,square, drools);
+            if (f == null) {
+                return;
+            }
 
             AddedNodeFact thumbNode = new AddedNodeFact(imageNode, name, "jnt:resource", drools);
             if (thumbNode.getNode() != null) {
@@ -144,13 +154,26 @@ public class ImageService {
         String savePath = org.jahia.settings.SettingsBean.getInstance().getTmpContentDiskPath();
         File Ftemp = new File(savePath);
         if (!Ftemp.exists()) Ftemp.mkdirs();
-        final File f = File.createTempFile("thumb","jpg", Ftemp);
+        String fileExtension = FilenameUtils.getExtension(imageNode.getName());
+        if ((fileExtension != null) && (!"".equals(fileExtension))) {
+            fileExtension += "." + fileExtension;
+        } else {
+            fileExtension = null;
+        }
+        final File f = File.createTempFile("thumb", fileExtension, Ftemp);
 
         Image iw = getImageWrapper(imageNode, drools);
+        if (iw == null) {
+            return null;
+        }
 
-        imageService.createThumb(iw, f, size, square);
-        f.deleteOnExit();
-        return f;
+        if (imageService.createThumb(iw, f, size, square)) {
+            f.deleteOnExit();
+            return f;
+        } else {
+            f.deleteOnExit();
+            return null;
+        }
     }
 
     public void setHeight(AddedNodeFact imageNode, String propertyName, KnowledgeHelper drools) throws Exception {
